refactor(utils): split path_helpers.py into focused modules
Split path_helpers.py into two focused modules:
- git_path_client.py: GitPathProtocol and Git wrapper functions
- path_helpers.py: Path manipulation utilities

Changes:
- Create git_path_client.py with GitPathProtocol (51 lines)
- Refactor path_helpers.py (493 → 433 lines, -60 lines)
- Update imports in 8 service files
- Update documentation in utils/README.md
- Remove dead code: BranchBoundGitClient (0 callers)
- Maintain backward compatibility via GitClientProtocol alias

Tests: 23 path_helpers + 317 service tests pass
Type check: mypy passes
Lint: ruff passes

Closes #532
