+++ config/loc_limits.yaml
@@ -141,7 +141,7 @@ code_limits:
       # Utils 层 —— 允许 500
       - path: "src/vibe3/utils/path_helpers.py"
         limit: 500
-        reason: "路径解析核心工具，包含 GitClientProtocol、路径相关 Git wrapper、BranchBoundGitClient 适配器和路径解析逻辑，虽然职责略显混杂但都围绕路径处理，暂不拆分"
+        reason: "路径解析核心工具，专注于路径解析和 handoff 路径处理。GitClientProtocol 和 Git wrapper 已移至 git_path_client.py，BranchBoundGitClient 已移除（死代码）"
 
       # 强耦合测试例外
       - path: "tests/vibe3/agents/test_codeagent_backend.py"
