+++ src/vibe3/services/flow_service.py
@@ -5,7 +5,7 @@ from vibe3.clients.git_client import GitClient
 from vibe3.config.settings import VibeConfig
 from vibe3.services.flow_block_mixin import FlowLifecycleMixin
 from vibe3.services.flow_transition import FlowTransitionMixin
-from vibe3.utils.path_helpers import GitClientProtocol
+from vibe3.utils.git_path_client import GitPathProtocol
 
 
 class FlowService(FlowLifecycleMixin, FlowTransitionMixin):
@@ -22,13 +22,13 @@ class FlowService(FlowLifecycleMixin, FlowTransitionMixin):
     """
 
     store: SQLiteClient
-    git_client: GitClientProtocol
+    git_client: GitPathProtocol
     config: VibeConfig
 
     def __init__(
         self,
         store: SQLiteClient | None = None,
-        git_client: GitClientProtocol | None = None,
+        git_client: GitPathProtocol | None = None,
         config: VibeConfig | None = None,
     ) -> None:
         """Initialize flow service.
